refactor(highlighter): highligher.ts to highlighter.ts
