+++ src/handler/refactor/buffer.ts
@@ -5,7 +5,7 @@ import { Position, Range, TextEdit } from 'vscode-languageserver-types'
 import { URI } from 'vscode-uri'
 import { createLogger } from '../../logger'
 import Document from '../../model/document'
-import Highlighter from '../../model/highligher'
+import Highlighter from '../../model/highlighter'
 import { DidChangeTextDocumentParams, Optional, TextDocumentContentChange } from '../../types'
 import { disposeAll } from '../../util'
 import { isParentFolder, readFileLines, sameFile } from '../../util/fs'
