+++ src/handler/refactor/search.ts
@@ -4,7 +4,7 @@ import { ChildProcess, spawn } from 'child_process'
 import { EventEmitter } from 'events'
 import { Range } from 'vscode-languageserver-types'
 import { createLogger } from '../../logger'
-import Highlighter from '../../model/highligher'
+import Highlighter from '../../model/highlighter'
 import { ansiparse } from '../../util/ansiparse'
 import { Mutex } from '../../util/mutex'
 import { path, readline } from '../../util/node'
