+++ src/handler/semanticTokens/index.ts
@@ -4,7 +4,7 @@ import commands from '../../commands'
 import events from '../../events'
 import languages from '../../languages'
 import BufferSync from '../../model/bufferSync'
-import Highlighter from '../../model/highligher'
+import Highlighter from '../../model/highlighter'
 import { Documentation, FloatFactory } from '../../types'
 import { disposeAll } from '../../util'
 import { distinct, isFalsyOrEmpty, toArray } from '../../util/array'
