+++ src/handler/workspace.ts
@@ -10,7 +10,7 @@ import { PatternType } from '../core/workspaceFolder'
 import extensions from '../extension'
 import languages, { ProviderName } from '../languages'
 import { getLoggerFile } from '../logger'
-import Highligher from '../model/highligher'
+import Highligher from '../model/highlighter'
 import snippetManager from '../snippets/manager'
 import { defaultValue } from '../util'
 import { CONFIG_FILE_NAME, isVim } from '../util/constants'
