+++ src/model/highlighter.ts
